Combine Mixer, Music and Audio, use the single mutex to avoid potential deadlocks, modernize code a bit
