+++ fheroes2-vs2019.vcxproj
@@ -174,8 +174,6 @@
   <ItemGroup>
     <ClCompile Include="src\engine\agg_file.cpp" />
     <ClCompile Include="src\engine\audio.cpp" />
-    <ClCompile Include="src\engine\audio_mixer.cpp" />
-    <ClCompile Include="src\engine\audio_music.cpp" />
     <ClCompile Include="src\engine\dir.cpp" />
     <ClCompile Include="src\engine\engine.cpp" />
     <ClCompile Include="src\engine\image.cpp" />
@@ -388,8 +386,6 @@
   <ItemGroup>
     <ClInclude Include="src\engine\agg_file.h" />
     <ClInclude Include="src\engine\audio.h" />
-    <ClInclude Include="src\engine\audio_mixer.h" />
-    <ClInclude Include="src\engine\audio_music.h" />
     <ClInclude Include="src\engine\dir.h" />
     <ClInclude Include="src\engine\engine.h" />
     <ClInclude Include="src\engine\image.h" />
