+++ src/fheroes2/battle/battle_arena.cpp
@@ -27,7 +27,7 @@
 #include "ai.h"
 #include "army.h"
 #include "army_troop.h"
-#include "audio_mixer.h"
+#include "audio.h"
 #include "battle_arena.h"
 #include "battle_army.h"
 #include "battle_bridge.h"
