+++ src/fheroes2/game/game_loadgame.cpp
@@ -23,7 +23,7 @@
 #include "game.h"
 #include "agg.h"
 #include "agg_image.h"
-#include "audio_mixer.h"
+#include "audio.h"
 #include "cursor.h"
 #include "dialog.h"
 #include "game_io.h"
