+++ src/fheroes2/game/game_mainmenu.cpp
@@ -22,7 +22,7 @@
 
 #include "agg.h"
 #include "agg_image.h"
-#include "audio_mixer.h"
+#include "audio.h"
 #include "cursor.h"
 #include "dialog.h"
 #include "dialog_language_selection.h"
