+++ src/fheroes2/game/game_newgame.cpp
@@ -23,8 +23,7 @@
 #include "game.h"
 #include "agg.h"
 #include "agg_image.h"
-#include "audio_mixer.h"
-#include "audio_music.h"
+#include "audio.h"
 #include "campaign_savedata.h"
 #include "cursor.h"
 #include "dialog.h"
