+++ src/fheroes2/game/game_video.cpp
@@ -19,7 +19,7 @@
  ***************************************************************************/
 
 #include "game_video.h"
-#include "audio_mixer.h"
+#include "audio.h"
 #include "cursor.h"
 #include "game_delays.h"
 #include "localevent.h"
@@ -97,7 +97,7 @@ namespace Video
 
         const uint32_t delay = static_cast<uint32_t>( 1000.0 / video.fps() + 0.5 ); // This might be not very accurate but it's the best we can have now
 
-        const bool hasSound = Mixer::isValid();
+        const bool hasSound = Audio::isValid();
         const std::vector<std::vector<uint8_t> > & sound = video.getAudioChannels();
         if ( hasSound ) {
             for ( std::vector<std::vector<uint8_t> >::const_iterator it = sound.begin(); it != sound.end(); ++it ) {
