+++ src/fheroes2/gui/interface_events.cpp
@@ -24,7 +24,7 @@
 #include <vector>
 
 #include "agg.h"
-#include "audio_mixer.h"
+#include "audio.h"
 #include "cursor.h"
 #include "dialog.h"
 #include "game.h"
