+++ src/tools/xmi2mid.vcxproj
@@ -171,8 +171,6 @@
   <ItemGroup>
     <ClCompile Include="..\engine\agg_file.cpp" />
     <ClCompile Include="..\engine\audio.cpp" />
-    <ClCompile Include="..\engine\audio_mixer.cpp" />
-    <ClCompile Include="..\engine\audio_music.cpp" />
     <ClCompile Include="..\engine\dir.cpp" />
     <ClCompile Include="..\engine\engine.cpp" />
     <ClCompile Include="..\engine\image.cpp" />
@@ -201,8 +199,6 @@
   <ItemGroup>
     <ClInclude Include="..\engine\agg_file.h" />
     <ClInclude Include="..\engine\audio.h" />
-    <ClInclude Include="..\engine\audio_mixer.h" />
-    <ClInclude Include="..\engine\audio_music.h" />
     <ClInclude Include="..\engine\dir.h" />
     <ClInclude Include="..\engine\engine.h" />
     <ClInclude Include="..\engine\image.h" />
